Moved spherical harmonics things from IBL.cpp to ShericalHarmonics class
